==> jupyter_notebooks/01 Data Collection load_pp_monthly.ipynb <==
"""
Script to load and filter the PPD data for the Midlands Counties only.

"""

import pandas as pd

csv_path = "../data/raw/pp-monthly-update-new-version.csv"

# read the CSV file, splitting on commas only if they are not inside double quotes.
df = pd.read_csv(
    csv_path,
    sep=r',(?=(?:[^"]*"[^"]*")*[^"]*$)',
    engine="python",
    header=0,
)
df.columns = df.columns.str.strip()

# clean string columns by trimming spaces and extra quotes
for col in df.select_dtypes(include="object"):
    df[col] = df[col].str.strip().str.strip(chr(34))

df.head()


# df shape
df.shape
df.info()
df.describe(include="all")

# check for missing values
df.isnull().sum()


df.duplicated().sum()



# Check the number of unique values in each column
df.nunique()

df.shape
# what are the lengths of the postcodes?
df["postcode"].str.len().value_counts().sort_index()

# drop rows with postcode null or length is 0
df = df[df["postcode"].notnull()]
df = df[df["postcode"].str.len() > 0]
df.shape

# focus on PPD category type A (standard sale) and status A (added)
df = df[(df["PPD_category"] == "A") & (df["Status"] == "A")]
df.shape

# # duplicates removal
# dups if postcode, PAON, SAON, street, locality, town, city, county are the same
# find these and display them
duplicates = df[df.duplicated(subset=["price", "postcode", "PAON", "SAON", "Street", "Locality", "town_city", "county"], keep=False)]
# order by postcode
duplicates = duplicates.sort_values(by=["postcode"])
duplicates

# remove duplicates, keeping the last occurrence
df = df.drop_duplicates(
    subset=["price", "postcode", "PAON", "SAON", "Street", "Locality", "town_city", "county"],
    keep="last"
)
df.shape

# Use these Midlands counties only
counties_to_use = [
    "STOKE-ON-TRENT",
    "CITY OF DERBY",
    "DERBYSHIRE",
    "WEST MIDLANDS",
    "STAFFORDSHIRE",
    "LEICESTERSHIRE",
    "LEICESTER",
    "WORCESTERSHIRE",
    "WARWICKSHIRE",
]

# filter to keep only these counties
df = df[df["county"].isin(counties_to_use)]
df.shape

df.to_csv("../data/clean/pp-monthly-cleaned.csv", index=False)
