==> jupyter_notebooks/02 Load and Join ONS Postcode Directory.ipynb <==
"""
02 Load and Join ONS Postcode Directory.
The primary purpose of this notebook is to facilitate spatial joins between property transaction data and geographic areas.
This allows for integration with other datasets such as the Index of Multiple Deprivation (IMD) and Energy Performance Certificates (EPC).
The analysis of housing prices and trends at different geographic levels within the Midlands is then possible.
Fields added: `lsoa11cd`, `msoa11nm`, `ladnm` to the Price Paid Data (PPD) dataset based on matching `postcode`.
Output: `ppd_with_geography.csv` - enriched PPD dataset with geographic identifiers.
"""
import pandas as pd

# Load the large PCD zipped CSV file with proper encoding
df = pd.read_csv('../data/raw/PCD_OA_LSOA_MSOA_LAD_MAY22_UK_LU.zip', compression='zip', encoding='latin1')

df.info()

df.head()

# Join with Price Paid Data (PPD) on the `postcode` field

# First load the previously cleaned Price Paid Data (PPD)
ppd = pd.read_csv('../data/clean/pp-monthly-cleaned.csv')


ppd.head()

# Perform the join
merged_df = pd.merge(ppd, df[['pcds', 'lsoa11cd', 'msoa11nm', 'ladnm']], left_on='postcode', right_on='pcds', how='left')
# The `pcds` column can now be dropped 
merged_df.drop(columns=['pcds'], inplace=True)
# Save the merged dataframe to a new CSV file
merged_df.to_csv('../data/clean/ppd_with_geography.csv', index=False)
