==> jupyter_notebooks/03 Load and Join IMD data.ipynb <==
"""
03 Load and Join IMD data
- Joins IMD data to the main dataframe using LSOA codes.
- fields: `lsoa11cd`, `imd_decile`, `imd_score`.
- Purpose: enrich the dataset with deprivation indices.
- Saves the enriched dataframe to a new CSV file: ppd_with_geography_and_imd.csv
"""
import pandas as pd

# Load the previously merged Price Paid Data (PPD) with ONS Postcode Directory
df = pd.read_csv('../data/clean/ppd_with_geography.csv')

# Load the IMD data
imd = pd.read_csv('../data/raw/imd2019lsoa.csv')


imd.head(50)


imd.info()

# unique values for Indices of Deprivation
imd['Indices of Deprivation'].unique()

In future a more granular approach may be done by individual indices rather than just the IMD index:
  - Index of Multiple Deprivation (IMD)
  - Income Deprivation Domain
  - Employment Deprivation Domain
  - Education, Skills and Training Domain
  - Health Deprivation and Disability Domain
  - Crime Domain
  - Barriers to Housing and Services Domain
  - Living Environment Deprivation Domain
  - Income Deprivation Affecting Children Index (IDACI)
  - Income Deprivation Affecting Older People Index (IDAOPI)

# Select relevant columns from IMD data where Measurement is Decile and Indices of Deprivation is 'a. Index of Multiple Deprivation (IMD)'
imd_decile = imd[(imd['Measurement'].str.strip() == 'Decile') &
                 (imd['Indices of Deprivation'].str.strip() == 'a. Index of Multiple Deprivation (IMD)')]
imd_rank = imd[(imd['Measurement'].str.strip() == 'Rank') &
               (imd['Indices of Deprivation'].str.strip() == 'a. Index of Multiple Deprivation (IMD)')]
# rename measurement columns
imd_decile = imd_decile.rename(columns={'Value': 'IMD_Decile'})
imd_rank = imd_rank.rename(columns={'Value': 'IMD_Rank'})
# combine decile and rank dataframes
imd_combined = pd.merge(imd_decile[['FeatureCode', 'IMD_Decile']], imd_rank[['FeatureCode', 'IMD_Rank']],
                        on='FeatureCode', how='inner')
imd_combined.head()


# Perform the join with the main dataframe
merged_df = pd.merge(df, imd_combined, left_on='lsoa11cd', right_on='FeatureCode', how='left')
# Drop the `FeatureCode` column as it's no longer needed
merged_df = merged_df.drop(columns=['FeatureCode'])
# check the result
merged_df.info()

# are there any missing IMD values?
merged_df['IMD_Decile'].isnull().sum(), merged_df['IMD_Rank'].isnull().sum()
# remove rows with missing IMD values
merged_df = merged_df.dropna(subset=['IMD_Decile', 'IMD_Rank'])
merged_df.info()

# Save the enriched dataframe to a new CSV file
merged_df.to_csv('../data/clean/ppd_with_geography_and_imd.csv', index=False)
